Everything works
Final commit to get everything working. Updates to expand functionality may come in the future, but this is to make everything work on the Yún rev.2. Had to update the original IRSendRev library to configure the pinouts from the ATMega32U4 to change the pinout to one supported by the Grove Shield